\name{mfvarr-package}
\alias{mfvarr-package}
\alias{mfvarr}
\docType{package}
\title{
\packageTitle{mfvarr}
}
\description{
\packageDescription{mfvarr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mfvarr}
\packageIndices{mfvarr}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{mfvarr}

Maintainer: \packageMaintainer{mfvarr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
